no apple comments

+++ s7_mark_files.py
@@ -1,11 +1,7 @@
 from json import loads
-from os import mkdir
-from os.path import isdir, join as path_join
-import applescript
-
-
-def set_comment(file_path, comment_text):
-    applescript.tell.app("Finder", f'set comment of (POSIX file "{file_path}" as alias) to "{comment_text}" as Unicode text')
+from os import mkdir, remove
+from os.path import isdir, join as path_join, isfile
+from exif import Image
 
 
 def mark_files(project_dir):
@@ -15,12 +11,24 @@ def mark_files(project_dir):
     with open(path_join(temp_dir, 'data.json'), 'r') as r:
         data_to_process = loads(r.read()) or {}
     downloaded_media = data_to_process['downloaded_media']
-    for path, meta in downloaded_media.items():
-        text = f"post: {meta['post_url']}\nauthor: {meta['author_url']}"
-        set_comment(path, text)
-        with open(path + '.txt', 'w') as w:
-            w.write(text)
+    total = len(downloaded_media)
+    for k, (path, meta) in enumerate(downloaded_media.items(), 1):
+        print('marking', k, 'of', total, '...')
+        text = f"p: {meta['post_url']}\na: {meta['author_url']}"
+        txt_path = path + '.txt'
+        if path.endswith('.jpg'):
+            try:
+                with open(path, 'rb') as image_file:
+                    my_image = Image(image_file)
+                my_image.user_comment = text
+                with open(path, 'wb') as new_image_file:
+                    new_image_file.write(my_image.get_file())
+            except Exception as e:
+                print('~~', 'Error', e, 'for file', path)
+        if not isfile(txt_path):
+            with open(path + '.txt', 'w') as w:
+                w.write(text)
 
 
 if __name__ == '__main__':
-    mark_files('/Volumes/Elements/inst_project')
+    mark_files('C:\\Users\\David\\PycharmProjects\\instagram-posts-parser')
